[Разбор Полетов] Remove weird `CTOC` and `CHAP` tags
They are causing exceptions when saving tags to a file and only warnings
when reading them. They look like new tags from apple.

As a sidenote, the stagger's API has two parts: the higher-level,
user-friendly API (like `album` and `date` properties) and the
lower-level, direct API to the tags, as in the standard. This fix is
possible due to the latter, which makes the library more flexible, yet
less safe.